Make Manager the main exported type
Splitting store and manager was a bit much, store is unlikely useful by itself
and the manager doesn't work fully for cookie (reset is useless)

Un-export the store and manager type, and have a distinct cookie and KV manager
with things that work for them. This will also make it a bit easier to massage
things around. In practice the two store types aren't really interchangable, if
implementations want it to be pluggable they can define their own interface.